+++ training/xtime/datasets/_gas_concentrations.py
@@ -73,7 +73,7 @@ class GasConcentrationsBuilder(DatasetBuilder):
         assert isinstance(y, pd.Series), f"Expecting y to be of type pd.Series (type = {type(y)})."
 
         # Encode labels (it's categorical column with values from [1, 6])
-        y = pd.Series(LabelEncoder().fit_transform(y), index=y.index, name=y.name)
+        y = pd.Series(LabelEncoder().fit_transform(y), index=y.index, name=y.name).astype("int32")
 
         # Remove instances with missing values
         # No need to remove instances with missing values - no such instances.
